desktop/src-tauri/python/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
pyyaml==6.0.1
psutil==5.9.5
mss==9.0.1
pytesseract==0.3.10
Pillow==10.0.0
apscheduler==3.10.4
tzlocal==5.2
pytz==2023.3
setuptools==69.0.3
six==1.16.0
